requirements.txt
easydict
easydl
tensorboardX
tqdm
pyyaml